chris.requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
configobj==5.0.6
ecdsa==0.16.0
idna==2.10
pbkdf2==1.3
pip-tools==5.3.1
pyaes==1.6.1
pyblake2==1.1.2
pycurl==7.43.0.6
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
slick-bitcoinrpc==0.1.4
ujson==3.2.0
urllib3==1.25.10
